Refactored response handling code out of BaseRequest into ResponseHandler class
