+++ src/Microsoft.Graph.Core/Requests/ResponseHandler.cs
@@ -1,0 +1,61 @@
+﻿// ------------------------------------------------------------------------------
+//  Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
+// ------------------------------------------------------------------------------
+
+namespace Microsoft.Graph
+{
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Net.Http;
+    using System.Threading.Tasks;
+
+    public class ResponseHandler
+    {
+        private readonly ISerializer serializer;
+
+        public ResponseHandler(ISerializer serializer)
+        {
+            this.serializer = serializer;
+        }
+
+        public async Task<T> HandleResponse<T>(HttpResponseMessage response)
+        {
+            if (response.Content != null)
+            {
+                var responseString = await GetResponseString(response);
+                return serializer.DeserializeObject<T>(responseString);
+            }
+
+            return default(T);
+        }
+
+        /// <summary>
+        /// Get the response content string
+        /// </summary>
+        /// <param name="hrm">The response object</param>
+        /// <returns>The full response string to return</returns>
+        private async Task<string> GetResponseString(HttpResponseMessage hrm)
+        {
+            var responseContent = "";
+
+            var content = await hrm.Content.ReadAsStringAsync().ConfigureAwait(false);
+
+            //Only add headers if we are going to return a response body
+            if (content.Length > 0)
+            {
+                var responseHeaders = hrm.Headers;
+                var statusCode = hrm.StatusCode;
+
+                Dictionary<string, string[]> headerDictionary = responseHeaders.ToDictionary(x => x.Key, x => x.Value.ToArray());
+                var responseHeaderString = serializer.SerializeObject(headerDictionary);
+
+                responseContent = content.Substring(0, content.Length - 1) + ", ";
+                responseContent += "\"responseHeaders\": " + responseHeaderString + ", ";
+                responseContent += "\"statusCode\": \"" + statusCode + "\"}";
+            }
+
+            return responseContent;
+        }
+
+    }
+}
